[hir2mpl]Internal commit msg:     [hir2mpl] add fe_err builtin_mpl_vector_st     [ICSL] Clean Code 6     [hir2mpl] bugfix atomic_load_n get type     [ICSL] Clean Code 7 [mapleall]Internal commit msg:     [ICSL] Clean Code 6     [intrinsic]Bugfix for DTS2023022000938     [driver]add new option     [clean_code 12]     [mplcg] tailcall opt enhance     [mplcg] add global-schedule & local-schedule phases     [mapleall] Do not print extra inline log if no dump option given     [mplcg] fix bug in x16 opt for getting memOpndIdx     [pgo]lite-pgo cfg bugfix     DTS2023022402157     [mapleall] cleancode2  16     [me] override IsSameContent in child classes for code factoring     [alias] set next level NADS for all mustdefs     [mplcg] replace all implicit cvt     [alias] correct offset of ost with fieldID in ddaa     [outline] deal stmt with OP_intrinsiccallwithtype as call-return nodes     [WIP][mplall] tailcall opt enhance

+++ src/mapleall/maple_me/include/me_value_range_prop.h
@@ -164,9 +164,9 @@ class Bound {
       return true;
     }
     if (IsPrimTypeUint64(fromType)) {
-      return static_cast<uint64>(constant) == GetRealValue(constant, toType);
+      return constant == GetRealValue(constant, toType);
     } else if (IsPrimTypeUint64(toType)) {
-      return static_cast<uint64>(constant) == GetRealValue(constant, fromType);
+      return constant == GetRealValue(constant, fromType);
     } else {
       return GetRealValue(constant, fromType) == GetRealValue(constant, toType);
     }
